the revolution begins
it's a mess.
